pyproject.toml
[project]
name = "allium-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
